Support aligned *vec* even when SIMD isn't enabled

+++ glm/detail/func_integer_simd.inl
@@ -57,7 +57,7 @@ namespace detail
 	template <>
 	GLM_FUNC_QUALIFIER int bitCount(uint64 x)
 	{
-		return _mm_popcnt_u64(x);
+		return static_cast<int>(_mm_popcnt_u64(x));
 	}
 #	endif
